Introduce an intermediate function to execute in a child process
Introducing an intermediate function to execute in a child process as we
wanted to make the usage of the decorator easy, and not change the
underlying function that is being decorated.

This intermediate function will take care of running the correct
function and appending the result value to a queue.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>