Refactor few functions in pkghandler
Refactored the way we handle package signatures and fingerprints during
the conversion as it was causing a lot of problems, with the conversion
(not being able to issue a Ctrl + C) and with the code itself.

This refactor moves away from using yum (dnf)/rpm APIs to use binary
calls to rpm to identify information about a package (or all installed
packages).

Since the function became more "generic", the `get_pkg_signature`
function went away and now, we have a `get_package_information` in
place.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>