Refactor remove_pkgs_with_confirm
We can't remove packages and modify state in a child process because
when we execute the rollback, we won't place back the packages that need
to be replaced on the system.

The function was split in two, one that get the packages (that runs in a
child process), and the original remove_pkgs_with_confirm, that does not
run in a child process and remove the packages.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>