Fix print_pkg_info with rpm installed package class
We used to use `hasattr` to figure out if the pkg being printed has the
attribute `nevra` in order to pass the correct param to the
`get_pkg_nevra` function. The problem with that approach is that
`hasattr` was turning the `pkg` object into a str, rather than the
RPMInstalledPackage instance that we need.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>